fix(database): a short primary key cannot hold 44292 BCards
BCard.BCardId is a short, and the parser writes more rows than a short can
index. Measured against a freshly parsed database: 44292 rows, highest id
66407, against a ceiling of 32767.

Two ways it shows, depending on how far the database got:

On a database still matching the migration, the column is smallint and the
insert fails once the identity passes 32767, so the parser cannot finish.

On a database whose column was widened to integer by hand to get past that,
the server starts and the WorldServer dies loading static data, because EF
still reads the column with GetInt16:

  System.OverflowException: Arithmetic operation resulted in an overflow.
     at Npgsql.Internal.Converters.Int4Converter`1.ReadCore(PgReader reader)
     at Npgsql.NpgsqlDataReader.GetInt16(Int32 ordinal)

That is the failure this was found on. It lands after "7688 Item loaded",
takes the Autofac activation down with it, and the process exits without
listening.

Nothing references BCardId as a foreign key, so widening it is contained to
the one column.

Tested: builds with 0 warnings, full suite green (991 tests). Verified on a
running server, which is what the fix is for: before, the WorldServer exited
during static load; after, it reaches "Listening on port 1337" and the
LoginServer registers against it.

+++ src/NosCore.Database/Entities/BCard.cs
@@ -12,7 +12,7 @@ namespace NosCore.Database.Entities
     public class BCard : IStaticEntity
     {
         [Key]
-        public short BCardId { get; set; }
+        public int BCardId { get; set; }
 
         public byte SubType { get; set; }
